Factor out the Home Assistant related logic to a separate file
